Big Update - First Complete Code Version
Unfortunately Github flaked out on me unexpectedly for much of the
trickier parts of development. I just managed to fix the issue so I can
commit again. A large number of additions have been made since the last
commit, and the code now seems to be working, although I have no way of
testing it to be certain. Plots of results are included in this folder
now.

Key additions since the last version:
- fixed units
- fixed Euler's method for both n,m,h and Vm
- realized Vm is equal to difference from resting potential
- included n0, m0, h0
- included plotting sections
- included excitation current input sections for each problem